requirements.txt
langchain-google-genai
langchain-chroma
langchain-community